requirements.txt
Flask>=2.3.2
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Limiter
Flask-Login
Flask-Mail
Flask-Bcrypt
Flask-PyMongo
gunicorn
dnspython
XlsxWriter
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
SQLAlchemy>=2.0.0
alembic
python-dotenv
Flask-Caching==2.0.2
redis
requests
PyMySQL
pymssql
pyodbc
cryptography
africastalking
Flask==2.3.2
Flask-DotEnv==0.1.2
flask-paginate==2024.4.12
pandas==2.2.3
openpyxl==3.1.5

# Optional, depending on your needs
schedule
rich
beautifulsoup4
